Current

+++ images/03.ServiceImages/backup/home/engines/scripts/engine/add_backup.sh
@@ -8,17 +8,24 @@ check_required_values
 backup_id=${parent}_${backup_name}_${src_type}
 dirname=${Backup_ConfigDir}/$backup_id
 
-echo -n $backup_type > $dirname/backup_type
+echo ADD Backup $1
+
+service_backup_cmd='curl -k https://172.17.0.1:2380/v0/backup' 
 
 echo dirname $dirname
 
-if test -d $dirname
-then 
- rm -rf $dirname
-fi
+if test -z $2
+ then
+   if test -d $dirname
+     then 
+      rm -rf $dirname
+   fi
+  mkdir -p $dirname
+fi  
 
-mkdir -p $dirname
 export dirname
+
+echo -n $backup_type > $dirname/backup_type
 
 if ! test -z "$email"
 then
@@ -33,12 +40,22 @@ else
  echo $backup_type >>/tmp/bt
  mkdir -p /tmp/backup
  src=/tmp/backup
- echo $1 >   $dirname/pre_cmd_path
- cat /home/engines/templates/backup/duply_pre > $dirname/pre
- cat /home/engines/templates/backup/duply_post > $dirname/post   
+ echo $1 > $dirname/pre_cmd_path
+ if ! test -f $dirname/pre
+  then
+   cat /home/engines/templates/backup/duply_pre > $dirname/pre
+   cat /home/engines/templates/backup/duply_post > $dirname/post
+  fi
+tp=`dirname $1`
+tp=`basename $tp`
+sh=`basename $1`
+backup_name=${backup_id}_$tp_$sh.backup
+  echo " $service_backup_cmd/engine/$1 > /tmp/backup/${backup_name} " >> $dirname/pre    
 fi
-
-write_duply_config
+if ! test -f $dirname/conf
+ then
+    write_duply_config
+fi    
 
 chmod u+x  $dirname/pre
 chmod u+x  $dirname/post
